feat(NFT) styled
