+++ src/components/NFTBalance/index.tsx
@@ -1,0 +1,2 @@
+export { default as NFTBalance } from './NFTBalance';
+export { INFTBalance } from './types';
