+++ src/components/NFTBalance/types.ts
@@ -1,0 +1,6 @@
+import { Chain } from '../../web3utils';
+
+export interface INFTBalance {
+    chain: Chain;
+    address: string;
+}
